Move update check into preValidate and avoid deep clone
The deep clone was because the update check was running in `validate`, after the op is actually applied, and so needed a copy of the old doc. However, the `preValidate` method has exactly the same semantics (contrary to the comment that was in this file previously about atomic ops), and runs before the op is applied, avoiding the need for this deep copy. It also bring it into alignment with the create and delete checks.